Documentation refactoring (#37)
* :pushpin:

Adding the Manifest file to enable full reproducibility, also
for updating most of the dependencies.

* :construction_worker:

Updated Julia version for the CI pipeline.

* :pushpin: :pencil:

Adding a different section for examples using Literate.jl

* :boom:

No need for the seperate markdown file.

* :construction: :pencil:

The documentation workflow for the examples is now ready, although
I'm still working on the documentation itself.

* :pencil: :sparkles:

Several more sections in the documentation, also some fixes in the
tests for the Simulated Annealing algorithm.

* :children_crossing:

Added a random seed API to all the solvers in order to ensure
reproducible results. Documentation updates are needed for these
changes.

* :pencil:

* :boom:

Added a seed parameter to the `Population` type in order to fully
ensure reproducibility.

* :pencil: :white_check_mark:

Had to remove the parallel implementation because it was not working
correctly due to the environments. I also updated the tests to include
checks for when a truly random seed is used.

* :pencil:

Removed demo from README as it is already implemented in the
documentation.

* :white_check_mark: