Documentation CI (#44)
* :construction_worker:

Working on the CI by fixing some parameters and paths.

* :construction_worker: :pencil:

This should take care of the branch, where I have changed it to the
stable version of the package. Updated the license information in the
full documentation and bumped a minor version of the package.

* :construction_worker:

No need to test on Windows 32-bit architecture.

* :construction_worker:

I was missing to explicitly name the development branch from which to
take the documentation for deployment.