Types refactoring (#46)
* :boom:

First step for the types refactoring involved streamlining the `Particle`
type to enforce type stability.

* :art:

This should take care of the type stability and refactoring.
Although is should be noted that the maximum and minimum for each particle
is not the right way to deal with boundaries. But for some reason, this
actually gives the right result in almost every benchmark.

* :art: :pencil:

Some more fixes to type annotations and docstrings. I have settled now
for the maximum and minimum for all searches. It turns out that specifying
bounds is a type of constrained optimization, which is not the purpose
of this PR.

* :boom:

There is no need for the `Individual` type, although it is a super-type,
we really aren't using it for anything.

* :pencil:

Even more work on the formatting of the documentation, but it should be
done for now.

* :pushpin:

Set the version to 0.7. Also, we didn't need Revise as an extra
dependency, that's just for development purposes.

* :white_check_mark:

Updated some tests as it turns out that we dropped in code coverage
on the last commits.