:art: :sparkles:
This is a big commit. First, the initial version of the `optimize`
interface has been correctly implemented. Right now it enable the dispatch
of PopulationBase solvers. Due to just having one of those, we cannot
test the full dispatch, but it should be possible.
Then, the passing of a single RNG has been implemented for both particles,
populations and solvers. The user must now always pass a custom RNG, or
leave it directly to the interface.