:sparkles: :pencil:
It took way too much time for the parallel example to run on the
GitHub Actions servers. Fortunately, Literate can execute the code
and capture the output. I have split the creation of the example and
the creation of the documentation. The examples are to be executed
in my computer, the result will be stored, and then the documentation
will be built. This will remove all need to run very long-running
codes on the servers.