Common interface implementation (#47)
* :art: :sparkles:

This is a big commit. First, the initial version of the `optimize`
interface has been correctly implemented. Right now it enable the dispatch
of PopulationBase solvers. Due to just having one of those, we cannot
test the full dispatch, but it should be possible.
Then, the passing of a single RNG has been implemented for both particles,
populations and solvers. The user must now always pass a custom RNG, or
leave it directly to the interface.

* :sparkles:

Added support for sorting Populations based on the best possible solution.

* :boom: :art:

Kind of a weird one. I attempted to implement the tolerance for convergence
but I guess these algorithms already have convergence built into them
such that iterations are only needed. Sometimes, the tolerance would make
the loop break because there were no changes within solutions, even
when those solutions were incorrect.

* :pencil:

Adding some documentation for the common optimization interface.

* :sparkles: :pencil:

Finished implementing the common interface for optimization, which will
hopefully dispatch correctly through every possible implementation.
Docstrings and some other documentation was written.
I had to introduce a new type, `TrajectoryBase` in order to dispatch
correctly to the SimulatedAnnealing implementations.

* :pencil:

The examples on the documentation needed refactoring to reflect
the changes of the new API. More work on the documentation is on the
way.

* :pencil:

Working on the documentation.

* :pencil: :art:

Some documentation restructuring. Added a considerable amount of information
on the no free lunch theorems.

* :pencil: :sparkles:

New example in the documentation that uses the newly implemented
`optimize` interface.
Some other fixes to the documentation, as well as enabling multithreading
in the CI pipeline.

* :pushpin: :construction_worker:

Dealing with some issues regarding CI.
Also, bumped the package to the v0.8 version.

* :sparkles: :pencil:

It took way too much time for the parallel example to run on the
GitHub Actions servers. Fortunately, Literate can execute the code
and capture the output. I have split the creation of the example and
the creation of the documentation. The examples are to be executed
in my computer, the result will be stored, and then the documentation
will be built. This will remove all need to run very long-running
codes on the servers.

* :pencil: :sparkles:

The examples are run beforehand, and the markdown files are now edited
before building the full documentation.

* :construction_worker:

No need to run the makedocs step multithreaded.

* :pencil:

Minor fix.