Bugfix (#49)
* :sparkles:

Instead of returning the bounds, generate a new random solution within
the same bounds.

* :sparkles: :art:

Added a new clipping function for trajectory based algorithms. Also did
some reformatting to stick to the Blue style.

* :sparkles:

Implemented the clipping in the Simulated Annealing implementations.

* :art:

Always return nothing to stick to Blue style.

* :bug:

No need to use the and operator.

* :fire:

It turns out that if you generate new solutions within the search space,
there is no improvement in performance. On the contrary, performance
worsens. I have opted to remove this implementation.

* :arrow_up: :pushpin:

All tests pass, and I have update to the latest dependencies.

* :white_check_mark: :art:

Some minor formatting. Added a new test to account for keyword
arguments. Everything seems to be working as expected.